+++ src/icons/ChevronLeft.jsx
@@ -1,12 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-const ChevronLeft = ({ height, ...other }) => (
+const ChevronLeft = ({ height, width, ...other }) => (
   <svg
     height={height}
+    width={width}
     viewBox="0 0 24 24"
-    version="1.1"
-    xmlns="http://www.w3.org/2000/svg"
     {...other}
   >
     <polygon points="11.1715729 12 15.4142136 16.2426407 14 17.6568542 8.34314575 12 14 6.34314575 15.4142136 7.75735931 11.1715729 12" />
@@ -15,10 +14,12 @@ const ChevronLeft = ({ height, ...other }) => (
 
 ChevronLeft.propTypes = {
   height: PropTypes.string,
+  width: PropTypes.string,
 };
 
 ChevronLeft.defaultProps = {
-  height: '2rem',
+  height: '1.5rem',
+  width: '1.5rem',
 };
 
 export default ChevronLeft;
